Add files via upload
This Excel file represents my first project in business-related data analysis. The dataset was created with hypothetical values for simplicity and easier modeling. However, it was processed using common and widely used Excel tools, and ultimately, it provides a concise management dashboard that summarizes the financial situation of a fictional company. In the future, I aim to expand my projects by utilizing more advanced tools and features in the field of data analysis.